pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain-community>=0.3.19",
    "langchain-openai>=0.3.8",
    "notebook>=7.3.3",
    "pydantic-ai>=0.0.39",
    "pypdf>=5.3.1",
]
